[FE] Fix: 데이터 2번 불러오는 문제 수정 #119
